Use weighted sampling for peer crawling
